store charging session into db

+++ backend/EMS.Library/Adapter/PriceProvider/Tariff.cs
@@ -4,10 +4,10 @@ namespace EMS.Library.Adapter.PriceProvider
 	public record Tariff
 	{
 		public DateTime Timestamp { get; }
-		public double TariffUsage { get; }
-		public double TariffReturn { get; }
+		public Decimal TariffUsage { get; }
+		public Decimal TariffReturn { get; }
 
-		public Tariff(DateTime timestamp, double tariffUsage, double tariffReturn)
+		public Tariff(DateTime timestamp, Decimal tariffUsage, Decimal tariffReturn)
 		{
 			Timestamp = timestamp;
 			TariffUsage = tariffUsage;
